backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
openai
python-dotenv
requests
pypdf2
python-docx
numpy
pgvector